feat(riscv): add v8 simulator foundation and math model verification
- v8 RISC-V layer: real skill tables from cnbe32.db, Python/C/QEMU/Verilog/Spike all PASS
- Linux 0.01 CNBE-32 RISC-V boots in QEMU/OpenSBI and aligns with v8 runtime (55/55)
- Math experiments: pseudo-metric boundary, distributive lattice + 3D range query, entropy analysis, hyperbolic geometry, algebraic property tests
- Add verify_project.sh unified local verification entry

+++ linux_cnbe32_riscv/lib/read.c
@@ -10,7 +10,7 @@
 #define __LIBRARY__
 #include <unistd.h>
 
-int read(int fd, char * buf, int count)
+int read(int fd, char * buf, off_t count)
 {
 	int __res;
 	register long a7 __asm__("a7") = __NR_read;
